backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pyodbc==5.0.1
pandas==2.1.3
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
